+++ src/main/java/ua/od/game/repository/RoomRepository.java
@@ -6,5 +6,5 @@ import ua.od.game.model.Room;
 /**
  * @author ruslan.gramatic on 6/27/18.
  */
-public interface RoomRepository extends JpaRepository<Room, Long> {
+public interface RoomRepository extends JpaRepository<Room, Integer> {
 }
